+++ internal/service/commit/commits_helper.go
@@ -68,7 +68,7 @@ func gitLog(ctx context.Context, sender lines.Sender, repo *pb.Repository, revis
 	if err != nil {
 		return err
 	}
-	defer cmd.Kill()
+	defer cmd.Close()
 
 	split := lines.ScanWithDelimiter([]byte("\x00"))
 	if err := lines.Send(cmd, sender, split); err != nil {