+++ CHANGELOG.md
@@ -4,6 +4,8 @@ UNRELEASED
 
 - Implement CommitService.FindCommits
   https://gitlab.com/gitlab-org/gitaly/merge_requests/266
+- Log spawned process metrics
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/284
 
 v0.30.0