+++ CHANGELOG.md
@@ -8,6 +8,8 @@ UNRELEASED
   https://gitlab.com/gitlab-org/gitaly/merge_requests/284
 - Implement RepositoryService.ApplyGitattributes RPC
   https://gitlab.com/gitlab-org/gitaly/merge_requests/278
+- Implement RepositoryService.FetchRemote RPC
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/276
 
 v0.30.0